% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_object.R
\name{view_object_addin}
\alias{view_object_addin}
\alias{view_object}
\title{View selected code in RStudio's viewer}
\usage{
view_object_addin()

view_object(text)
}
\description{
Send the selected text to \code{View()} in Rstudio.
}
